Crate winrt_notification

An incomplete wrapper over the WinRT toast api

Tested in Windows 10 and 8.1. Untested in Windows 8, might work.

Todo:

Add support for Adaptive Content

Add support for Actions

Known Issues:

Will not work for Windows 7.

Will not build when targeting the 32-bit gnu toolchain (i686-pc-windows-gnu).

Limitations:

Windows 8.1 only supports a single image, the last image (icon, hero, image) will be the one on the toast

Structs

Error
A WinRT error object consists of both an error code as well as detailed error information for debugging.

Toast

Enums

Duration

IconCrop

LoopableSound
Sounds suitable for Looping

Sound