Struct winrt_notification::Toast

pub struct Toast { /* fields omitted */ }

Methods

impl Toast

fn new(app_id: &str) -> Toast

Constructor for the toast builder

app_id is used in the toast to inform the user where it came from.

fn title(self, content: &str) -> Toast

Sets the title of the toast.

Will be white. Supports Unicode ✓

fn text1(self, content: &str) -> Toast

Add/Sets the first line of text below title.

Will be grey. Supports Unicode ✓

fn text2(self, content: &str) -> Toast

Add/Sets the second line of text below title.

Will be grey. Supports Unicode ✓

fn duration(self, duration: Duration) -> Toast

Set the length of time to show the toast

fn icon(self, source: &Path, crop: IconCrop, alt_text: &str) -> Toast

Set the icon shown in the upper left of the toast

The default is supposed to be determined by your app id. In practice it will be blank.

fn hero(self, source: &Path, alt_text: &str) -> Toast

Add/Set a Hero image for the toast.

This will be above the toast text and the icon.

fn image(self, source: &Path, alt_text: &str) -> Toast

Add an image to the toast

May be done many times. Will appear below text.

fn sound(self, src: Option<Sound>) -> Toast

Set the sound for the toast or silence it

Default is Sound::IM

fn show(&self) -> Result<(), Error>

Display the toast on the screen